@@ -12,17 +12,18 @@
     }
   ],
   "require": {
-    "php": ">=5.4.0",
-    "illuminate/console": "5.*",
-    "illuminate/database": "5.*",
-    "illuminate/events": "5.*",
-    "illuminate/filesystem": "5.*",
-    "illuminate/support": "5.*"
+    "php": "^7.2",
+    "illuminate/console": "^6.0",
+    "illuminate/database": "^6.0",
+    "illuminate/events": "^6.0",
+    "illuminate/filesystem": "^6.0",
+    "illuminate/support": "^6.0",
+    "laravel/helpers": "^1.2"
   },
   "require-dev": {
-    "phpunit/phpunit": "~4.0",
-    "mockery/mockery": "~0.9",
-    "d11wtq/boris": "~1.0.10"
+    "phpunit/phpunit": "^7.0",
+    "mockery/mockery": "^1.3",
+    "d11wtq/boris": "^1.0"
   },
   "autoload": {
     "psr-0": {
