@@ -3,17 +3,20 @@
 class CategorySoftDeletesTest extends CategoryTestCase {
 
   public function testReload() {
+      /**
+       * @var \Baum\Node $node
+       */
     $node = $this->categories('Child 3', 'SoftCategory');
 
     $node->delete();
 
     $this->assertTrue($node->trashed());
-    $this->assertFalse($node->exists);
+    $this->assertNotEmpty($node->deleted_at);
 
-    $node->reload();
+    $node->restore();
 
-    $this->assertTrue($node->trashed());
-    $this->assertTrue($node->exists);
+    $this->assertFalse($node->trashed());
+    $this->assertEmpty($node->deleted_at);
   }
 
   public function testDeleteMaintainsTreeValid() {
@@ -243,7 +246,7 @@ class CategorySoftDeletesTest extends CategoryTestCase {
   public function testAllStatic() {
     $expected = array('Root 1', 'Child 1', 'Child 2', 'Child 2.1', 'Child 3', 'Root 2');
 
-    $this->assertArraysAreEqual($expected, SoftCategory::all()->lists('name'));
+    $this->assertArraysAreEqual($expected, SoftCategory::all()->pluck('name')->toArray());
   }
 
   public function testAllStaticWithSoftDeletes() {
@@ -251,7 +254,7 @@ class CategorySoftDeletesTest extends CategoryTestCase {
     $this->categories('Child 3', 'SoftCategory')->delete();
 
     $expected = array('Root 1', 'Child 2', 'Child 2.1', 'Root 2');
-    $this->assertArraysAreEqual($expected, SoftCategory::all()->lists('name'));
+    $this->assertArraysAreEqual($expected, SoftCategory::all()->pluck('name')->toArray());
   }
 
 }
