@@ -24,8 +24,8 @@ class ClusterHierarchyTest extends ClusterTestCase {
     $this->assertEquals($query->count(), $roots->count());
     $this->assertCount(2, $roots);
 
-    foreach ($query->lists('id') as $node)
-      $this->assertContains($node, $roots->lists('id'));
+    foreach ($query->pluck('id') as $node)
+      $this->assertContains($node, $roots->pluck('id'));
   }
 
   public function testRootsStaticWithCustomOrder() {
@@ -35,7 +35,7 @@ class ClusterHierarchyTest extends ClusterTestCase {
     $roots = OrderedCluster::roots()->get();
 
     $this->assertCount(3, $roots);
-    $this->assertEquals($cluster, $roots->first());
+    $this->assertTrue($cluster->is($roots->first()));
   }
 
   public function testRootStatic() {
@@ -47,7 +47,7 @@ class ClusterHierarchyTest extends ClusterTestCase {
 
     $this->assertCount(4, $allLeaves);
 
-    $leaves = $allLeaves->lists('name');
+    $leaves = $allLeaves->pluck('name');
 
     $this->assertContains('Child 1'   , $leaves);
     $this->assertContains('Child 2.1' , $leaves);
@@ -60,7 +60,7 @@ class ClusterHierarchyTest extends ClusterTestCase {
 
     $this->assertCount(1, $allTrunks);
 
-    $trunks = $allTrunks->lists('name');
+    $trunks = $allTrunks->pluck('name');
     $this->assertContains('Child 2', $trunks);
   }
 
@@ -188,16 +188,16 @@ class ClusterHierarchyTest extends ClusterTestCase {
 
     $node = $this->clusters('Child 2');
 
-    $descendancy = $node->descendants()->lists('id');
+    $descendancy = $node->descendants()->pluck('id')->toArray();
 
-    $this->assertEmpty($node->descendants()->limitDepth(0)->lists('id'));
+    $this->assertEmpty($node->descendants()->limitDepth(0)->pluck('id'));
     $this->assertEquals($node, $node->descendantsAndSelf()->limitDepth(0)->first());
 
-    $this->assertEquals(array_slice($descendancy, 0, 3), $node->descendants()->limitDepth(3)->lists('id'));
-    $this->assertEquals(array_slice($descendancy, 0, 5), $node->descendants()->limitDepth(5)->lists('id'));
-    $this->assertEquals(array_slice($descendancy, 0, 7), $node->descendants()->limitDepth(7)->lists('id'));
+    $this->assertEquals(array_slice($descendancy, 0, 3), $node->descendants()->limitDepth(3)->pluck('id')->toArray());
+    $this->assertEquals(array_slice($descendancy, 0, 5), $node->descendants()->limitDepth(5)->pluck('id')->toArray());
+    $this->assertEquals(array_slice($descendancy, 0, 7), $node->descendants()->limitDepth(7)->pluck('id')->toArray());
 
-    $this->assertEquals($descendancy, $node->descendants()->limitDepth(1000)->lists('id'));
+    $this->assertEquals($descendancy, $node->descendants()->limitDepth(1000)->pluck('id')->toArray());
   }
 
   public function testGetAncestorsAndSelf() {
